Long overdue update time (#29)
Adds a Pale Troll Mini boss to the bottom level of the map with an ancient blade of strife as a reward.

Adds a Cavern Crab mob to the game to act as an angry source of food, they gather near water to lay their eggs.

Removes 3 fortresses for player density to be increased, from 6 to 3 fortresses.